requirements.txt
paho-mqtt==1.6.1
dronekit==2.9.2